Update train_loop_rl.py
Updated the whole training loop for RL, some of the changes are: 1) The BOCPD update function is updated so that the shift appears here, 2) It always starts from the first step now, instead of ignoring the samples of size "state_window", which is removed now. 3) Adding the best_params as parameters, so when hyperparameters are tuned, they can be directly passed to this function.