Update distance_spread.py
Creating a cointegration spread. Currently, distance and cointegration spreads are identical. The reason is that we want to rename the distance spread to cointegration spread, but in order to avoid backward compatibility for the previous, we will keep it this way for a few days and then remove distance spread or change distance spread code to match its functionality if required.